feat: add test case for repair bom, mop settings to the workflow
